xxx
git-svn-id: svn://elaine/murooma/trunk@69 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ timeUtils.h
@@ -26,7 +26,7 @@ std::string chunkTime(const T& chunk)
 }
 
 
-int diff_ms(const timeval& t1, const timeval& t2)
+long diff_ms(const timeval& t1, const timeval& t2)
 {
     return (((t1.tv_sec - t2.tv_sec) * 1000000) + 
             (t1.tv_usec - t2.tv_usec))/1000;
@@ -34,7 +34,7 @@ int diff_ms(const timeval& t1, const timeval& t2)
 
 
 template <typename T>
-int getAge(const T& chunk)
+long getAge(const T& chunk)
 {
 	timeval now;
 	gettimeofday(&now, NULL);
